+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (IN PROGRESS)
 
+* Previous fiscal year selected by default when viewing group records. Refs UIF-419.
+
 ## [3.3.0](https://github.com/folio-org/ui-finance/tree/v3.3.0) (2022-10-25)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v3.2.1...v3.3.0)